Add smart ai feature to targeting
